Downgraded to python2.7. Now it works. Pending porting to python3.
